move bokeh code out of diagnostics
This isolates bokeh code, reduces import time, and generally separates
concerns.
