+++ distributed/bokeh/tests/test_resource_monitor.py
@@ -7,7 +7,7 @@ from bokeh.plotting import Figure
 
 from tornado import gen
 from distributed import Executor
-from distributed.diagnostics.resource_monitor import Occupancy
+from distributed.bokeh.resource_monitor import Occupancy
 from distributed.utils_test import cluster, loop
 from time import time
