+++ distributed/bokeh/tests/test_status_monitor.py
@@ -1,4 +1,4 @@
-from distributed.diagnostics.status_monitor import (worker_table_plot,
+from distributed.bokeh.status_monitor import (worker_table_plot,
         worker_table_update, task_table_plot, task_table_update,
         task_stream_plot, task_stream_append, progress_plot)
 from distributed.diagnostics.scheduler import workers, tasks
