+++ distributed/bokeh/worker_monitor.py
