+++ distributed/cli/dscheduler.py
@@ -24,6 +24,11 @@ import signal
 
 
 def handle_signal(sig, frame):
+    try:
+        import distributed.bokeh
+        distributed.bokeh.server_process.kill()
+    except Exception:
+        pass
     IOLoop.instance().add_callback(IOLoop.instance().stop)
 
 signal.signal(signal.SIGINT, handle_signal)
@@ -52,11 +57,11 @@ def main(center, host, port, http_port, bokeh_port, show, _bokeh):
     if _bokeh:
         try:
             import bokeh
-            import distributed.diagnostics.bokeh
+            import distributed.bokeh
             hosts = ['%s:%d' % (h, bokeh_port) for h in
                      ['localhost', '127.0.0.1', ip, socket.gethostname(), host]]
             dirname = os.path.dirname(distributed.__file__)
-            paths = [os.path.join(dirname, 'diagnostics', 'bokeh', name)
+            paths = [os.path.join(dirname, 'bokeh', name)
                      for name in ['status', 'tasks']]
             args = (['bokeh', 'serve'] + paths +
                     ['--log-level', 'warning',
@@ -68,7 +73,7 @@ def main(center, host, port, http_port, bokeh_port, show, _bokeh):
                 args.append('--show')
             proc = subprocess.Popen(args)
 
-            distributed.diagnostics.bokeh.server_process = proc  # monkey patch
+            distributed.bokeh.server_process = proc  # monkey patch
 
             logger.info(" Start Bokeh UI at:        http://%s:%d/status/"
                         % (ip, bokeh_port))
@@ -80,6 +85,11 @@ def main(center, host, port, http_port, bokeh_port, show, _bokeh):
     loop.start()
     loop.close()
     scheduler.stop()
+    if _bokeh:
+        try:
+            distributed.bokeh.server_process.kill()
+        except:
+            pass
 
     logger.info("End scheduler at %s:%d", ip, port)
