+++ setup.py
@@ -22,9 +22,9 @@ setup(name='distributed',
       packages=['distributed',
                 'distributed.cli',
                 'distributed.diagnostics',
-                'distributed.diagnostics.bokeh',
-                'distributed.diagnostics.bokeh.status',
-                'distributed.diagnostics.bokeh.tasks',
+                'distributed.bokeh',
+                'distributed.bokeh.status',
+                'distributed.bokeh.tasks',
                 'distributed.http',
                 'distributed.formats'],
       long_description=(open('README.md').read() if os.path.exists('README.md')
